image-filter-web/requirements.txt
Flask
Pillow
numpy
gunicorn
